fix: resolve table a11y, contrast, reduced motion, and responsive mobile layout
