+++ tailwind.config.ts
@@ -39,7 +39,7 @@ const config: Config = {
           300: "#dee2e6",
           400: "#ced4da",
           500: "#adb5bd",
-          600: "#868e96",
+          600: "#9ca3af",
           700: "#495057",
           800: "#343a40",
           900: "#212529",
